# openapi.yaml
openapi: 3.1.0
info:
  title: Leave Management System API
  version: 1.0.0
  description: |
    A comprehensive API for managing employee leave requests, work-from-home requests,
    holiday planning, and document generation. This API supports role-based access control
    with different permissions for Employees, Managers, HR, Executives, and Admins.

    ## Authentication
    All endpoints (except `/api/health`) require authentication via NextAuth JWT sessions.
    Authentication is handled via Azure AD OAuth 2.0.

    ## Authorization
    API endpoints are protected by role-based access control (RBAC):
    - **EMPLOYEE**: Basic access to own requests and team calendar
    - **MANAGER**: Can approve team member requests
    - **DEPARTMENT_DIRECTOR**: Department-level oversight
    - **HR**: User management, document verification
    - **EXECUTIVE**: Company-wide analytics and approvals
    - **ADMIN**: Full system configuration access

  contact:
    name: API Support
    email: support@company.com
  license:
    name: Proprietary
    url: https://company.com/license
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://leave.company.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/hr/document-verification:
    get:
      tags:
        - HR
      summary: Get documents for verification
      description: Returns sick leave documents requiring HR verification
      operationId: getDocumentsForVerification
      responses:
        "200":
          description: Documents requiring verification
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/DocumentVerification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/executive/pending-approvals:
    get:
      tags:
        - Executive
      summary: Get executive pending approvals
      description: Returns requests awaiting executive approval (peer approvals, escalated requests)
      operationId: getExecutivePendingApprovals
      responses:
        "200":
          description: Pending approvals
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeaveRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/admin/departments:
    get:
      tags:
        - Admin
      summary: List departments
      operationId: getDepartments
      responses:
        "200":
          description: Department list
          content:
            application/json:
              schema:
                type: object
                properties:
                  departments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Department"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - Admin
      summary: Create department
      operationId: createDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "200":
          description: Department created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/holiday-planning/overlaps:
    get:
      tags:
        - Holiday Planning
      summary: Get holiday overlaps
      description: Detect overlapping holiday plans in team
      operationId: getHolidayOverlaps
      responses:
        "200":
          description: Overlap information
          content:
            application/json:
              schema:
                type: object
                properties:
                  overlaps:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        users:
                          type: array
                          items:
                            type: string
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Notifications
  /api/notifications:
    get:
      tags:
        - Notifications
      summary: Get notifications
      operationId: getNotifications
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  unreadCount:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/setup/auth:
    post:
      tags:
        - Setup
      summary: Authenticate for setup
      description: Authenticates with SETUP_PASSWORD to access setup endpoints
      operationId: setupAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Authentication successful
        "401":
          description: Invalid password
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth JWT session token
    cronSecret:
      type: http
      scheme: bearer
      description: CRON_SECRET environment variable for scheduled tasks
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Forbidden
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
    RequestStatus:
      type: string
      enum: [DRAFT, PENDING, APPROVED, REJECTED, CANCELLED]
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    DocumentStatus:
      type: string
      enum: [DRAFT, PENDING_SIGNATURES, COMPLETED]
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: string
    LeaveType:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        daysAllowed:
          type: integer
        carryForward:
          type: boolean
        maxCarryForward:
          type: integer
          nullable: true
        requiresApproval:
          type: boolean
        requiresDocument:
          type: boolean
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        maxDaysPerRequest:
          type: integer
          nullable: true
        isSpecialLeave:
          type: boolean
        requiresHRVerification:
          type: boolean
        documentTypes:
          type: array
          items:
            type: string
...
    LeaveRequest:
      type: object
      properties:
        id:
          type: string
        requestNumber:
          type: string
        userId:
          type: string
        leaveTypeId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        totalDays:
          type: number
        reason:
          type: string
        substituteId:
          type: string
          nullable: true
        selectedDates:
          type: array
          items:
            type: string
            format: date-time
        status:
          $ref: "#/components/schemas/RequestStatus"
        documentUrl:
          type: string
          nullable: true
        supportingDocuments:
          type: object
          nullable: true
        hrDocumentVerified:
          type: boolean
        hrVerifiedBy:
          type: string
          nullable: true
        hrVerifiedAt:
          type: string
          format: date-time
          nullable: true
        hrVerificationNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        leaveType:
          $ref: "#/components/schemas/LeaveType"
        user:
          $ref: "#/components/schemas/UserSummary"
        substitute:
          $ref: "#/components/schemas/UserSummary"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/Approval"
        generatedDocument:
          $ref: "#/components/schemas/GeneratedDocument"
    Approval:
      type: object
      properties:
        id:
          type: string
        leaveRequestId:
          type: string
        approverId:
          type: string
        level:
          type: integer
        status:
          $ref: "#/components/schemas/ApprovalStatus"
        comments:
          type: string
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        signature:
          type: string
          nullable: true
        signedAt:
          type: string
          format: date-time
          nullable: true
        escalatedToId:
          type: string
          nullable: true
        escalatedAt:
          type: string
          format: date-time
          nullable: true
        escalationReason:
          type: string
          nullable: true
        approver:
          $ref: "#/components/schemas/UserSummary"
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        order:
          type: integer
...
    GeneratedDocument:
      type: object
      properties:
        id:
          type: string
        templateId:
          type: string
        leaveRequestId:
          type: string
        fileUrl:
          type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        templateSnapshot:
          type: object
        decisions:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        signatures:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSignature"
    DocumentSignature:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        signerId:
          type: string
        signerRole:
          type: string
        signatureData:
          type: string
        signedAt:
          type: string
          format: date-time
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [LEAVE_REQUESTED, LEAVE_APPROVED, LEAVE_REJECTED, LEAVE_CANCELLED, APPROVAL_REQUIRED, DOCUMENT_READY]
        title:
          type: string
        message:
          type: string
        link:
          type: string
          nullable: true
        isRead:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    DocumentVerification:
      type: object
      properties:
        id:
          type: string
        leaveRequestId:
          type: string
        requestNumber:
          type: string
        employeeName:
          type: string
        leaveType:
          type: string
        documents:
          type: array
          items:
            type: string
        submittedAt:
          type: string
          format: date-time
...
    CreateDepartmentBody:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        order:
          type: integer
...